srcs/Frontend/requirements.txt
dash==3.1.0
numpy==2.3.1
pandas==2.3.0
plotly==6.1.2
python-dotenv==1.1.1
Requests==2.32.4
openpyxl==3.1.5
